+++ src/main/java/com/kakao/termproject/pet/dto/PetCreateRequest.java
@@ -4,6 +4,7 @@ import com.kakao.termproject.pet.domain.Gender;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public record PetCreateRequest(
 
@@ -33,7 +34,7 @@ public record PetCreateRequest(
 
   @NotNull
   @Schema(description = "몸무게", example = "4")
-  float weight,
+  BigDecimal weight,
 
   @Schema(description = "선호 날씨", example = "sun")
   String preferredWeather,
